requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1605055780383/work
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.4
Django @ file:///tmp/build/80754af9/django_1606860386887/work
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
praw==7.1.0
prawcore==1.5.0
psycopg2 @ file:///C:/ci/psycopg2_1594305172475/work
pylint==2.6.0
python-dotenv==0.15.0
pytz @ file:///tmp/build/80754af9/pytz_1606604771399/work
requests==2.25.0
six==1.15.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
tabulate==0.8.7
toml==0.10.2
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
wincertstore==0.2
wrapt==1.12.1
yapf==0.30.0
